+++ HISTORY.md
@@ -1,3 +1,6 @@
+- wspot v1.0.2 changelog:
+1. [installer:] fixed auto updater.
+2. [poster:] fixed latest tag release upload.
 - wspot v1.0.1 changelog:
 1. [installer] added restore legacy database.
 2. [installer] fixed a issue on ufetch function.